# [controls] improve perf of "merged" ResourceDictionary lookups